requested changes / fixes / refactorings

+++ Habitica/src/main/java/com/habitrpg/android/habitica/models/user/HabitRPGUser.java
@@ -220,8 +220,8 @@ public class HabitRPGUser extends BaseModel {
         return this.balance;
     }
 
-    public double getGemCount(){
-        return this.balance * 4;
+    public int getGemCount(){
+        return (int)(this.balance * 4);
     }
 
     public void setBalance(double balance) {
